+++ EssentialFeed/EssentialFeedTests/Feed Cache/LoadFeedImageDataFromCacheUseCaseTests.swift
@@ -85,16 +85,6 @@ class LoadFeedImageDataFromCacheUseCaseTests: XCTestCase {
         XCTAssertTrue(receivedResults.isEmpty)
     }
     
-    func test_saveImageDataForURL_requestsImageDataSaveForURL() {
-        let (sut, store) = makeSUT()
-        let imageData = anyData()
-        let url = anyURL()
-        
-        sut.save(image: imageData, for: url) { _ in }
-        
-        XCTAssertEqual(store.messages, [.insert(data: imageData, for: url)])
-    }
-    
     // MARK: - Helpers
     
     private func makeSUT(currentDate: @escaping (() -> Date) = Date.init, file: StaticString = #filePath, line: UInt = #line) -> (sut: LocalFeedImageDataLoader, store: FeedImageDataStoreSpy) {
